Changed instance mapping implementation with jackson deserialization usage
