+++ src/test/resources/instances/bulk/expectedInstance.json
@@ -143,13 +143,5 @@
   },
   "natureOfContentTermIds": [
     "96879b60-098b-453b-bf9a-c47866f1ab2a"
-  ],
-  "precedingTitles": [
-    {
-      "title": "Houston oil directory"
-    }
-  ],
-  "succeedingTitles": [],
-  "parentInstances": [],
-  "childInstances": []
+  ]
 }
